Quick Sort
QuickSort is a Divide and Conquer algorithm. It picks an element as pivot and partitions the given array around the picked pivot. There are many different versions of quickSort that pick pivot in different ways. 
Always pick first element as pivot.
Always pick last element as pivot (implemented in code)
Pick a random element as pivot.