Graph Color (Mapping)
PROBLEM STATEMENT: 

Each of the states is divided into many districts for administrative reasons. These districts have many neighbors that share their borders. If we want to colour each of the districts is such a manner that adjoint districts (districts that share border) cannot be coloured with the same colour, perform for the map with customized border sharing but at least 8 or more districts.