Matrix Word Search
PROBLEM STATEMENT: 
Given a 2D board of characters and a word to search, to find if the word exists in the grid. The word can be constructed from letters of sequentially adjacent cells, where adjacent cells are those horizontally or vertically neighboring. The same letter cell may not be used more than once.