BackTracking
Maze Problem: Given set of location of points where if there is a way to go through or not. if taken a wrong turn (to a dead end) backtrack to the path where turn was made and continue further. Position starting from source and ending at desired location (maze[n-1][n-1]), Directions of movements: Forward and Downwards ( no diagonals )